Use cross platform openURLWithHelper
