+++ shared/settings/invites/container.js
@@ -1,8 +1,8 @@
 // @flow
 import React, {Component} from 'react'
-import {ipcRenderer} from 'electron'
 import Invites from './index'
 import {invitesReclaim, invitesRefresh, invitesSend, notificationsSave, notificationsToggle} from '../../actions/settings'
+import {openURLWithHelper} from '../../util/open-url'
 
 import {navigateAppend} from '../../actions/route-tree'
 
@@ -38,7 +38,7 @@ export default connector.connect(
       onSave: () => { dispatch(notificationsSave()) },
       onToggle: (name: string) => dispatch(notificationsToggle(name)),
       onToggleUnsubscribeAll: () => dispatch(notificationsToggle()),
-      onSelectUser: (username: string) => { ipcRenderer.send('openURL', 'user', {username}) },
+      onSelectUser: (username: string) => { openURLWithHelper('user', {username}) },
       onSelectPendingInvite: (invite: PendingInvite) => {
         dispatch(navigateAppend([{
           selected: 'inviteSent',
