+++ shared/util/open-url.js.flow
@@ -1,3 +1,4 @@
 // @flow
-export default function openURL (url: ?string) : void {}
+export default function openURL (url: ?string): void {}
+export function openURLWithHelper (type: string, params: ?string): void {}
